[CustomOp] start refactoring custom op sys, move hl ops to general
