+++ src/finn/custom_op/general/maxpoolnhwc.py
