+++ src/finn/custom_op/general/streamingdataflowpartition.py
