+++ src/finn/custom_op/general/xnorpopcount.py
